See description
Api validation checks if obj ref props reference a real obj in the DB, but there is a delay between when a new record is created and when it can be found via search. That means if you create a record that references a record that was just created, there's a chance validation would not find the connected record and would throw an error. I have been trying to account for this by having the api require an array of newly created objRefProps, but now I think that might be unnecessary. I need to rethink the logic, make a decision and code accordingly.