complex prob see desc
NewInputWSearchUniqueNew takes 2 actions: 1st it calls the API service to check if the new Name already exists, then it calls the onUpdateProp method to update the WMP state w the new value. The problem is onUpdateProp calls csValidateObj which ALSO searches for a matching name, but csValidateObj is calling an old url and not going through the new apiService. But even if we fix what csValidateObj calls, we still have a duplicate server call happening. Need to clean this up...