Complex prob, see desc, cont.
Per last commit desc - valService is searching for Name with old url and returning error. The issue is this valError findByName fetch is looking for something different than the NewInputWSearchUniqueNew component. The later is only checking the DB if the name of the current obj has changed, and then it is looking for ANY match, to prevent a non-unique name. The former is checking for a list of matching records, and is then checking if one of those records === this obj. I should change logic so both cases work like the later so I can handle the reqs with one route.