Fix for real-time mac tbl changes, other issues persist
Changes to the WMP now reflect to the Macro Tables immediately. Now trying to turn "Just Created" styling on/off appropriately, depending on what was created and whether the created record was changed. Some machinery is in place, but Save Changes needs work to handle when you create an Ingredient while editing a just-created RecipeIngredient.